Attempt to fix build issues
