+++ packages/desktop-client/src/components/HelpMenu.tsx
@@ -9,6 +9,7 @@ import { useToggle } from 'usehooks-ts';
 import { pushModal } from 'loot-core/client/actions/modals';
 
 import { SvgHelp } from '../icons/v2/Help';
+import { getPageDocs } from '../util/help';
 import { openUrl } from '../util/router-tools';
 
 import { Button } from './common/Button2';
@@ -44,26 +45,6 @@ const HelpButton = forwardRef<HTMLButtonElement, HelpButtonProps>(
 );
 
 HelpButton.displayName = 'HelpButton';
-
-export const getPageDocs = (page: string) => {
-  switch (page) {
-    case '/budget':
-      return 'https://actualbudget.org/docs/getting-started/envelope-budgeting';
-    case '/reports':
-      return 'https://actualbudget.org/docs/reports/';
-    case '/schedules':
-      return 'https://actualbudget.org/docs/schedules';
-    case '/payees':
-      return 'https://actualbudget.org/docs/transactions/payees';
-    case '/rules':
-      return 'https://actualbudget.org/docs/budgeting/rules';
-    case '/settings':
-      return 'https://actualbudget.org/docs/settings';
-    default:
-      // All pages under /accounts, plus any missing pages
-      return 'https://actualbudget.org/docs';
-  }
-};
 
 export const HelpMenu = () => {
   const { t } = useTranslation();
